Fix ranker node bug.

+++ app/inners/models/value_objects/contracts/requests/passage_searchs/process_body.py
@@ -18,6 +18,6 @@ class ProcessBody(BaseRequest):
     embedding_dimension: int
     num_iterations: Optional[int]
     similarity_function: str
-    retriever_top_k: float
-    ranker_top_k: float
+    retriever_top_k: int
+    ranker_top_k: int
     api_key: Optional[str]
